Inotify apenas no cliente
