+++ include/server/server.h
@@ -14,10 +14,6 @@
 
 struct clientList *clientList; //Inicialização do nodo inicial da lista de clientes
 
-/*
-Lança uma thread para ficar no watcher no path de argumento
-*/
-void *inotifyWatcher(void *pathToWatch);
 
 void *handleConnection(void *socketDescriptor);
 /*
@@ -29,10 +25,6 @@ void appendNewClient(int socketNewClient, char* userName);
   colocar em option INSERTDEVICE(0) para inserir e REMOVE DEVICE(1) para remover
 */
 int updateNumberOfDevices(struct clientList *client_node, int socketNumber, int option);
-/*
-  Verifica se o usuario já tem o diretorio criado na pasta User com seu nome, e se não tiver já o cria
-*/
-int checkAndCreateDir(char *pathName);
 
 struct client
 {
