+++ src/server/server.c
@@ -14,64 +14,6 @@
 #include "../../include/common/common.h"
 #include "../../include/linkedlist/linkedlist.h"
 
-void *inotifyWatcher(void *pathToWatch){
-    int length;
-    int fd;
-    int wd;
-    char buffer[BUF_LEN];
-
-    fd = inotify_init();
-
-        if ( fd < 0 ) {
-        perror( "inotify_init" );
-    }
-
-    wd = inotify_add_watch( fd, (char *) pathToWatch, 
-                            IN_CLOSE_WRITE | IN_CREATE | IN_DELETE | IN_MOVED_FROM | IN_MOVED_TO);
-
-    while (1) {
-        int i = 0;
-        length = read( fd, buffer, BUF_LEN );
-
-        if ( length < 0 ) {
-            perror( "read" );
-        }  
-
-        while ( i < length ) {
-            struct inotify_event *event = ( struct inotify_event * ) &buffer[ i ];
-            if ( event->len ) {
-                if ( event->mask & IN_CREATE ) {
-                    if ( event->mask & IN_ISDIR ) {
-                        printf( "The directory %s was created in %s.\n", event->name,(char *) pathToWatch);       
-                    }
-                    else {
-                        printf( "The file %s was created in %s.\n", event->name,(char *) pathToWatch);
-                    }
-                }
-                else if ( event->mask & IN_DELETE ) {
-                    if ( event->mask & IN_ISDIR ) {
-                        printf( "The directory %s was deleted in %s.\n", event->name,(char *) pathToWatch);       
-                    }
-                    else {
-                        printf( "The file %s was deleted in %s.\n", event->name,(char *) pathToWatch);
-                    }
-                }
-                else if ( event->mask & IN_MODIFY ) {
-                    if ( event->mask & IN_ISDIR ) {
-                        printf( "The directory %s was modified in %s.\n", event->name,(char *) pathToWatch );
-                    }
-                    else {
-                        printf( "The file %s was modified in %s.\n", event->name,(char *) pathToWatch);
-                    }
-                }
-            }
-            i += EVENT_SIZE + event->len;
-        }
-    }
-    ( void ) inotify_rm_watch( fd, wd );
-    ( void ) close( fd );
-}
-
 void *handleConnection(void *socketDescriptor) {
     char buffer[PACKET_SIZE];
     int exitCommand = FALSE;
@@ -80,20 +22,13 @@ void *handleConnection(void *socketDescriptor) {
     int idUserName;
     char *userName = malloc(sizeof(userName));
     char pathServerUsers[30] = "";
-    char pathServerToClient[30] = "../client/";
     pthread_t thread_id;
 
-    //TODO: Receives username from client
-
+  
     //TODO: get_sync_dir, creates directory, if not created
 
-    //TODO: Create here new thread to watch folder
-    // Inotify?
-    //
 
     //TODO: Thread to receive updates from client
-
-    //TODO: Linked list to link logged in clients
 
     /*************************************/
     //Reads the client name and update/search on the client list.
@@ -103,25 +38,19 @@ void *handleConnection(void *socketDescriptor) {
         printf("ERROR reading from socket");
     strcpy(userName , buffer);
     strcat(pathServerUsers,buffer);
-    strcat(pathServerToClient,buffer);
 
     struct clientList *client_node = malloc(sizeof(*client_node));//node used to find the username on the list.
 
     if(!findNode(buffer, clientList, &client_node)){
         appendNewClient(newsockfd, buffer);
         /*
-        Cria se necessario a pasta do usuario no cliente e/ou no servidor
+        Cria se necessario a pasta do usuario no servidor
         */
         checkAndCreateDir(pathServerUsers);
-        checkAndCreateDir(pathServerToClient);
         /*
-        Lança as thread que fica olhando o diretorio do cliente e do servidor
+        Lança a thread que fica olhando o diretorio do servidor
         */
         if(pthread_create(&thread_id, NULL, inotifyWatcher, (void *) pathServerUsers) < 0){
-			    fprintf(stderr,"ERROR, could not create thread.\n");
-			    exit(-1);
-		}
-        if(pthread_create(&thread_id, NULL, inotifyWatcher, (void *) pathServerToClient) < 0){
 			    fprintf(stderr,"ERROR, could not create thread.\n");
 			    exit(-1);
 		}
@@ -135,18 +64,13 @@ void *handleConnection(void *socketDescriptor) {
         }
         else{
             /*
-            Cria se necessario a pasta do usuario no cliente e/ou no servidor
+            Cria se necessario a pasta do usuario no servidor
             */
             checkAndCreateDir(pathServerUsers);
-            checkAndCreateDir(pathServerToClient);
             /*
-            Lança as thread que fica olhando o diretorio do cliente e do servidor
+            Lança a thread que fica olhando o diretorio do servidor
             */
             if(pthread_create(&thread_id, NULL, inotifyWatcher, (void *) pathServerUsers) < 0){
-			    fprintf(stderr,"ERROR, could not create thread.\n");
-			    exit(-1);
-		    }
-            if(pthread_create(&thread_id, NULL, inotifyWatcher, (void *) pathServerToClient) < 0){
 			    fprintf(stderr,"ERROR, could not create thread.\n");
 			    exit(-1);
 		    }
@@ -259,22 +183,3 @@ int updateNumberOfDevices(struct clientList *client_node, int socketNumber, int 
     return 0;
 }
 
-int checkAndCreateDir(char *pathName){
-    struct stat sb;
-    //printf("%s",strcat(pathcomplete, userName));
-    if (stat(pathName, &sb) == 0 && S_ISDIR(sb.st_mode)){
-        // usuário já tem o diretório com o seu nome
-        return 0;
-    }
-    else{
-        if (mkdir(pathName, 0777) < 0){
-            //....
-            return -1;
-        }
-        // diretório não existe
-        else{
-            printf("Creating %s directory...\n", pathName);
-            return 0;
-        }
-    }
-}
